+++ CHANGES.md
@@ -2,6 +2,7 @@
 
 - Added Shift-G key binding to go to the end of text in Output-window.
 - Added 'cls' command to clear Output-window.
+- Added cursorline to Pods window.
 
 ## 0.1